requirements.txt
matplotlib
numpy
oauth2client
Pyrebase4
scikit-image